+++ src/packages/shared-types/action-types/index.ts
@@ -2,4 +2,4 @@ export * from "./toggle-withdraw-rai-enabled";
 export * from "./issue-rai";
 export * from "./respond-to-rai";
 export * from "./withdraw-rai";
-export * from "./withdrawPackage";
+export * from "./withdraw-package";
